# data/openapi.yaml
openapi: 3.0.0
info:
  title: Lead Ingestion API
  description: Comprehensive lead management and ingestion system API
  version: 1.0.0
  contact:
    name: Lead Management Team
    email: api-support@example.com
  license:
    name: Apache 2.0
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /assignments/{lead_id}:
    get:
      summary: Get lead assignment
      description: Retrieve current assignment for a lead
      tags:
        - Assignments
      operationId: getAssignment
      parameters:
        - name: lead_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Assignment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "404":
          description: Assignment not found

    post:
      summary: Reassign lead
      description: Manually assign or reassign a lead to a team member
      tags:
        - Assignments
      operationId: reassignLead
      parameters:
        - name: lead_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  format: uuid
                reason:
                  type: string
              required:
                - user_id
      responses:
        "200":
          description: Lead reassigned successfully
        "400":
          description: Invalid assignment
        "404":
          description: Lead or user not found
...
components:
  schemas:
...
    Assignment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lead_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        assigned_at:
          type: string
          format: date-time
        reassigned_at:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token for API authentication
security:
  - bearerAuth: []